Flutter V0.1 (#92)
* Wrap with scaffold to handle background

* reformat code

* create block model

* create Blocks Serializer and convert json to usable dart block models

* initiate block builder from serializer

* create the new node builder and add grid loader with node loader cuibtre

* change the dropdown list width to match all values

* refactor

* change block categorize location

* automate input and output nodes

* meeting edits

* change in backend response from input_dots to input_channels and add the display_name to the blockModel

* add the category and index to the block model and add exception handling

* update the blocks scheme

* refactor serializer and make it more readable

* seperate server call from serializer

* change backgroundColor,UI Apperance and add reusable button

* add final block nodes appearence

* debug the floating bar location misplaced

* create the model creation template output function

* refactor category builder

* add listener to user changes in model parameters

* create general input and output data class and extend it to model input an output class\n and add params to json getter to the blockmodel

* error handling in api calls and create preprocessor node interface

* refactor project path

* refactor project path

* refactor server project path

* remove server manager from gitignore

* remove server manager from git ignore

* update the run server path

* create interfaces folder and start working on data loader

* working serialization for models, fitter and data_loader

* debugged output function and create a working serializer

* refactor node builder and rename block to node

* add displayed name

* added tooltip to the node

* update and delete node in server and application

* debug delete nodes

* ignore this commit 'database update'

* added description to the components

* ignore pyc files

* refactor the apicall class and add description to the node

* untrack db

* rename node_server_calls and apply DI and refactor the node server calls

* updated DB to postgress 16

---------

+++ project/ai_operations/models.py
@@ -8,7 +8,7 @@ from django.db import models
 #     updated_at = models.DateTimeField(auto_now=True)
 
 class Node(models.Model):
-    node_id = models.IntegerField(primary_key=True)
+    node_id = models.BigIntegerField(primary_key=True)
     node_name = models.CharField(max_length=255)
     message = models.CharField(max_length=255, default="Done")
     node_data = models.BinaryField(null=True, blank=True)
@@ -25,7 +25,7 @@ class Node(models.Model):
 class Component(models.Model):
     displayed_name = models.CharField(max_length=255, default="")
     description = models.CharField(max_length=255, default="")
-    idx = models.IntegerField(default=0)
+    idx = models.BigIntegerField(default=0)
     category = models.CharField(max_length=255, default="")
     node_name = models.CharField(max_length=255)
     node_type = models.CharField(max_length=255, default="general")
